+++ src/uw_scan/storage/repository.py
@@ -38,6 +38,7 @@ from ._helpers import (
     status_family_for,
 )
 from .audit import _AuditMixin
+from .flow import _aggressor_label_confidence, _flow_footprint_label, _FlowMixin
 from .health import _HealthMixin
 from .jobs import _JobsMixin
 from .market_data import _MarketDataMixin
@@ -264,36 +265,6 @@ def _oi_change_bias(rows: list[dict[str, Any]]) -> str | None:
     return "mixed"
 
 
-def _flow_footprint_label(row: models.FlowAlert) -> str:
-    ask = row.total_ask_side_prem or Decimal(0)
-    bid = row.total_bid_side_prem or Decimal(0)
-    total = row.total_premium or ask + bid
-    ask_ratio = ask / total if total and total > 0 else None
-    if row.has_sweep and ask_ratio is not None and ask_ratio >= Decimal("0.65"):
-        return "directional_whale"
-    if row.has_multileg:
-        return "hedge_flow"
-    if row.has_floor:
-        return "dealer_hedge"
-    if ask_ratio is not None and Decimal("0.40") <= ask_ratio <= Decimal("0.60"):
-        return "gamma_scalper"
-    return "unclassified"
-
-
-def _aggressor_label_confidence(row: models.FlowAlert) -> Decimal | None:
-    ask = row.total_ask_side_prem or Decimal(0)
-    bid = row.total_bid_side_prem or Decimal(0)
-    total = row.total_premium or ask + bid
-    if total is None or total <= 0:
-        return None
-    dominant_side_ratio = max(ask, bid) / total
-    structure_penalty = Decimal("0.05") if row.has_multileg else Decimal(0)
-    confidence = min(
-        Decimal("1"), max(Decimal("0"), dominant_side_ratio - structure_penalty)
-    )
-    return confidence.quantize(Decimal("0.01"))
-
-
 def _vrp_sign_flip_status_for_db(value: bool | str | None) -> str | None:
     if value is True:
         return "true"
@@ -318,6 +289,7 @@ logger = logging.getLogger(__name__)
 
 class Repository(
     _AuditMixin,
+    _FlowMixin,
     _HealthMixin,
     _JobsMixin,
     _MarketDataMixin,
@@ -715,148 +687,8 @@ class Repository(
             for row in rows
         ]
 
-    # ------------------------------------------------------------------
-    # flow_events
-    # ------------------------------------------------------------------
-    def insert_flow_events(
-        self, run_id: int, ticker: str, alerts: Iterable[models.FlowAlert]
-    ) -> int:
-        rows = list(alerts)
-        if not rows:
-            return 0
-        sql = (
-            f"INSERT INTO {self._schema}.flow_events ("
-            "run_id, alert_id, ticker, option_chain, expiry, strike, option_type, "
-            "price, underlying_price, total_size, total_premium, "
-            "total_ask_side_prem, total_bid_side_prem, volume, open_interest, "
-            "volume_oi_ratio, has_sweep, has_floor, has_multileg, "
-            "all_opening_trades, iv_start, iv_end, alert_rule, "
-            "flow_footprint_label, aggressor_label_confidence, "
-            "rule_id, sector, issue_type, next_earnings_date, created_at) VALUES ("
-            "%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, "
-            "%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) "
-            "ON CONFLICT (run_id, alert_id) DO NOTHING"
-        )
-        with self._conn.cursor() as cur:
-            for r in rows:
-                cur.execute(
-                    sql,
-                    (
-                        run_id,
-                        r.id,
-                        r.ticker,
-                        r.option_chain,
-                        r.expiry,
-                        r.strike,
-                        r.type,
-                        r.price,
-                        r.underlying_price,
-                        r.total_size,
-                        r.total_premium,
-                        r.total_ask_side_prem,
-                        r.total_bid_side_prem,
-                        r.volume,
-                        r.open_interest,
-                        r.volume_oi_ratio,
-                        r.has_sweep,
-                        r.has_floor,
-                        r.has_multileg,
-                        r.all_opening_trades,
-                        r.iv_start,
-                        r.iv_end,
-                        r.alert_rule,
-                        r.flow_footprint_label or _flow_footprint_label(r),
-                        r.aggressor_label_confidence
-                        if r.aggressor_label_confidence is not None
-                        else _aggressor_label_confidence(r),
-                        r.rule_id,
-                        r.sector,
-                        r.issue_type,
-                        r.next_earnings_date,
-                        r.created_at,
-                    ),
-                )
-        return len(rows)
-
-    def upsert_flow_alerts_daily_rollup(
-        self,
-        *,
-        run_id: int,
-        ticker: str,
-        alerts: Iterable[models.FlowAlert],
-        alert_limit: int,
-        trade_date: _date | None = None,
-    ) -> None:
-        rows = list(alerts)
-        if trade_date is None:
-            trade_date = self._flow_alert_trade_date(rows)
-
-        bull_premium = Decimal("0")
-        bear_premium = Decimal("0")
-        ask_side_premium = Decimal("0")
-        bid_side_premium = Decimal("0")
-        total_premium = Decimal("0")
-        rules: Counter[str] = Counter()
-
-        for row in rows:
-            premium = row.total_premium or Decimal("0")
-            total_premium += premium
-            opt_type = (row.type or "").lower()
-            if opt_type == "call":
-                bull_premium += premium
-            elif opt_type == "put":
-                bear_premium += premium
-            ask_side_premium += row.total_ask_side_prem or Decimal("0")
-            bid_side_premium += row.total_bid_side_prem or Decimal("0")
-            if row.alert_rule:
-                rules[row.alert_rule] += 1
-
-        top_alert_rule = rules.most_common(1)[0][0] if rules else None
-
-        sql = (
-            f"INSERT INTO {self._schema}.flow_alerts_daily_rollup ("
-            "ticker, trade_date, run_id, alert_count, alert_count_is_limited, "
-            "total_premium, bull_premium, bear_premium, ask_side_premium, "
-            "bid_side_premium, top_alert_rule) "
-            "VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) "
-            "ON CONFLICT (ticker, trade_date) DO UPDATE SET "
-            "run_id=EXCLUDED.run_id, alert_count=EXCLUDED.alert_count, "
-            "alert_count_is_limited=EXCLUDED.alert_count_is_limited, "
-            "total_premium=EXCLUDED.total_premium, "
-            "bull_premium=EXCLUDED.bull_premium, "
-            "bear_premium=EXCLUDED.bear_premium, "
-            "ask_side_premium=EXCLUDED.ask_side_premium, "
-            "bid_side_premium=EXCLUDED.bid_side_premium, "
-            "top_alert_rule=EXCLUDED.top_alert_rule, updated_at=now()"
-        )
-        with self._conn.cursor() as cur:
-            cur.execute(
-                sql,
-                (
-                    ticker.upper(),
-                    trade_date,
-                    run_id,
-                    len(rows),
-                    len(rows) >= alert_limit,
-                    total_premium,
-                    bull_premium,
-                    bear_premium,
-                    ask_side_premium,
-                    bid_side_premium,
-                    top_alert_rule,
-                ),
-            )
-
-    def _flow_alert_trade_date(self, rows: list[models.FlowAlert]) -> _date:
-        market_tz = ZoneInfo("America/New_York")
-        for row in rows:
-            if row.created_at is None:
-                continue
-            created_at = row.created_at
-            if created_at.tzinfo is None:
-                created_at = created_at.replace(tzinfo=market_tz)
-            return created_at.astimezone(market_tz).date()
-        return datetime.now(market_tz).date()
+    # flow_events + flow_alerts_daily_rollup methods moved to _FlowMixin
+    # _flow_alert_trade_date is now module-level in flow.py (doesn't use self)
 
     # ------------------------------------------------------------------
     # Time-series history (UPSERT by (ticker, market_date))
